# openapi/v2/mgmt/openapi.bundle.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "0.14.0",
    "title": "Management API",
    "contact": {
      "name": "Supaglue",
      "email": "docs@supaglue.com",
      "url": "https://supaglue.com"
    },
    "description": "#### Introduction\n\nWelcome to the Management API documentation. You can use this API to manage customer integrations and connections.\n\n[![Run in Postman](https://run.pstmn.io/button.svg)](https://www.postman.com/supaglue/workspace/supaglue-public/overview)\n\n#### Base API URL\n\n```\nhttps://api.supaglue.io/mgmt/v2\n```\n"
  },
  "servers": [
    {
      "url": "https://api.supaglue.io/mgmt/v2",
      "description": "Supaglue API"
    }
  ],
  "paths": {
...
    "/magic_links": {
      "get": {
        "operationId": "getMagicLinks",
        "tags": [
          "Magic Links"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "List magic links",
        "description": "Get a list of magic links",
        "responses": {
          "200": {
            "description": "Magic Links",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/magic_link"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMagicLink",
        "summary": "Create magic link",
        "tags": [
          "Magic Links"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create_magic_link"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entity created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/magic_link"
                }
              }
            }
          }
        }
      }
    },
    "/magic_links/{magic_link_id}": {
      "delete": {
        "operationId": "deleteMagicLink",
        "summary": "Delete magic link",
        "tags": [
          "Magic Links"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "responses": {
          "204": {
            "description": "Magic Link deleted"
          }
        }
      },
      "parameters": [
        {
          "name": "magic_link_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": "6bf954f9-dba3-4c34-9275-3830dccb0030"
          }
        }
      ]
    },
...
    "/sync_configs": {
      "get": {
        "operationId": "getSyncConfigs",
        "tags": [
          "SyncConfigs"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "List Sync Configs",
        "description": "Get a list of Sync Configs",
        "responses": {
          "200": {
            "description": "SyncConfigs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/sync_config"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createSyncConfig",
        "summary": "Create Sync Config",
        "tags": [
          "SyncConfigs"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create_update_sync_config"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SyncConfig created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sync_config"
                }
              }
            }
          }
        }
      }
    },
    "/sync_configs/{sync_config_id}": {
      "get": {
        "operationId": "getSyncConfig",
        "summary": "Get Sync Config",
        "tags": [
          "SyncConfigs"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "responses": {
          "200": {
            "description": "SyncConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sync_config"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateSyncConfig",
        "summary": "Update Sync Config",
        "tags": [
          "SyncConfigs"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create_update_sync_config"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SyncConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sync_config"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteSyncConfig",
        "summary": "Delete Sync Config",
        "tags": [
          "SyncConfigs"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "responses": {
          "200": {
            "description": "SyncConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sync_config"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "sync_config_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "example": "d583ec72-55d4-4f35-9668-f6c759674cc1"
          }
        }
      ]
    },
...
    "/syncs": {
      "get": {
        "operationId": "getSyncs",
        "tags": [
          "Syncs"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Get Syncs",
        "description": "Get a list of Syncs.",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw"
            },
            "description": "The pagination cursor value"
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "100"
            },
            "description": "Number of results to return per page"
          },
          {
            "name": "customer_id",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "my-customer-1"
            },
            "description": "The customer ID that uniquely identifies the customer in your application"
          },
          {
            "name": "provider_name",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "salesforce"
            },
            "description": "The provider name"
          },
          {
            "name": "object_type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "common",
                "standard",
                "custom"
              ],
              "example": "standard"
            },
            "description": "The object type to filter by"
          },
          {
            "name": "object",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "contact"
            },
            "description": "The object to filter by"
          },
          {
            "name": "entity_id",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "e74b5a4f-f252-4a6c-940c-f35873498543"
            },
            "description": "The entity id to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "Sync",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/pagination"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/sync_with_provider_and_customer"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "API key to allow developers to access the API"
      }
    },
    "schemas": {
      "pagination": {
        "type": "object",
        "properties": {
          "next": {
            "type": "string",
            "nullable": true,
            "example": "eyJpZCI6IjQyNTc5ZjczLTg1MjQtNDU3MC05YjY3LWVjYmQ3MDJjNmIxNCIsInJldmVyc2UiOmZhbHNlfQ=="
          },
          "previous": {
            "type": "string",
            "nullable": true,
            "example": "eyJpZCI6IjBjZDhmYmZkLWU5NmQtNDEwZC05ZjQxLWIwMjU1YjdmNGI4NyIsInJldmVyc2UiOnRydWV9"
          },
          "total_count": {
            "type": "number",
            "example": 100
          }
        },
        "required": [
          "next",
          "previous",
          "total_count"
        ]
      },
...
      "magic_link": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ec208408-db29-4705-b39e-4d33070b4ef6"
          },
          "status": {
            "type": "string",
            "example": "created"
          },
          "application_id": {
            "type": "string",
            "example": "d8ceb3ff-8b7f-4fa7-b8de-849292f6ca69"
          },
          "customer_id": {
            "type": "string",
            "example": "my-customer-1"
          },
          "provider_id": {
            "type": "string",
            "example": "677fcfca-cf89-4387-a189-71c885be67bc"
          },
          "provider_name": {
            "$ref": "#/components/schemas/provider_name"
          },
          "url": {
            "type": "string",
            "description": "The magic link URL.",
            "example": "https://app.hubspot.com/contacts/123456"
          },
          "return_url": {
            "type": "string",
            "description": "URL to redirect to after the connection is authorized.",
            "example": "https://app.myapp.com/connections/123456"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-02-23T00:00:00Z"
          }
        },
        "required": [
          "id",
          "status",
          "application_id",
          "customer_id",
          "provider_id",
          "provider_name",
          "expires_at",
          "url",
          "return_url"
        ]
      },
      "create_magic_link": {
        "type": "object",
        "properties": {
          "customer_id": {
            "type": "string",
            "example": "my-customer-1"
          },
          "provider_name": {
            "$ref": "#/components/schemas/provider_name"
          },
          "return_url": {
            "type": "string",
            "description": "URL to redirect to after the connection is authorized.",
            "example": "https://app.myapp.com/connections/123456"
          },
          "expiration_secs": {
            "type": "number",
            "example": 18000
          }
        },
        "required": [
          "customer_id",
          "provider_name",
          "expiration_secs",
          "return_url"
        ]
      },
...
      "sync_config": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "465fdcb7-26b4-4090-894c-67cab41022bb"
          },
          "application_id": {
            "type": "string",
            "example": "9572d08b-f19f-48cc-a992-1eb7031d3f6a"
          },
          "destination_id": {
            "type": "string",
            "example": "6e7baa88-84dd-4dbc-902a-14522c2984eb"
          },
          "provider_id": {
            "type": "string",
            "example": "7f72ec07-e5c1-47fd-8cf5-e71dd13873af"
          },
          "config": {
            "$ref": "#/components/schemas/sync_config_data"
          }
        },
        "required": [
          "id",
          "application_id",
          "destination_id",
          "provider_id",
          "config"
        ]
      },
      "sync_config_data": {
        "type": "object",
        "properties": {
          "default_config": {
            "type": "object",
            "description": "Configuration options for \"how\" to sync.",
            "properties": {
              "period_ms": {
                "type": "integer",
                "minimum": 60000,
                "example": 60000,
                "description": "The period (in milliseconds) to run the sync."
              },
              "strategy": {
                "type": "string",
                "enum": [
                  "full then incremental",
                  "full only"
                ],
                "description": "\\\"full then incremental\\\" will run a full sync on the first run, then incremental syncs. \\\"full only\\\" will run a full sync on every run."
              },
              "auto_start_on_connection": {
                "type": "boolean",
                "description": "If true, the sync will start automatically when the connection is created.\nDefault: true\n"
              }
            },
            "required": [
              "period_ms",
              "strategy"
            ]
          },
          "common_objects": {
            "description": "A list of Supaglue objects to be synced.",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "object": {
                  "type": "string",
                  "description": "The Supaglue object name",
                  "example": "contact"
                }
              },
              "required": [
                "object"
              ]
            }
          },
          "standard_objects": {
            "description": "A list of case-sensitive Provider objects to be synced.",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "object": {
                  "type": "string",
                  "description": "The Provider object name (case sensitive)",
                  "example": "Contact"
                }
              },
              "required": [
                "object"
              ]
            }
          },
          "entities": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "entity_id": {
                  "type": "string",
                  "example": "3a82409f-c98f-4d25-bbd8-3335de3f12cc"
                }
              },
              "required": [
                "entity_id"
              ]
            }
          }
        },
        "required": [
          "default_config"
        ]
      },
...
      "provider_name": {
        "type": "string",
        "enum": [
          "hubspot",
          "salesforce",
          "pipedrive",
          "zendesk_sell",
          "ms_dynamics_365_sales",
          "zoho_crm",
          "capsule",
          "outreach",
          "gong",
          "apollo",
          "salesloft",
          "intercom",
          "linear",
          "clearbit",
          "6sense",
          "marketo",
          "salesforce_marketing_cloud_account_engagement"
        ],
        "example": "hubspot"
      },
...
      "sync_with_provider_and_customer": {
        "oneOf": [
          {
            "type": "object",
            "title": "Data model: Entities",
            "properties": {
              "id": {
                "type": "string",
                "example": "971cb76d-9558-42fe-8f3b-8a531c32bd5f"
              },
              "type": {
                "type": "string",
                "enum": [
                  "entity"
                ],
                "example": "entity"
              },
              "entity_id": {
                "type": "string",
                "example": "70da3830-b717-49cb-91f0-22fb3f09a97a"
              },
              "connection_id": {
                "type": "string",
                "example": "3217ea51-11c8-43c9-9547-6f197e02e5e4"
              },
              "sync_config_id": {
                "type": "string",
                "example": "3217ea51-11c8-43c9-9547-6f197e02e5e5"
              },
              "paused": {
                "type": "boolean",
                "example": false
              },
              "provider_name": {
                "type": "string",
                "example": "hubspot"
              },
              "customer_id": {
                "type": "string",
                "example": "my-customer-1"
              }
            },
            "required": [
              "id",
              "type",
              "entity_id",
              "connection_id",
              "sync_config_id",
              "paused",
              "provider_name",
              "customer_id"
            ]
          },
          {
            "type": "object",
            "title": "Data model: Objects",
            "properties": {
              "id": {
                "type": "string",
                "example": "971cb76d-9558-42fe-8f3b-8a531c32bd5f"
              },
              "type": {
                "type": "string",
                "enum": [
                  "object"
                ],
                "example": "object"
              },
              "object_type": {
                "type": "string",
                "enum": [
                  "common",
                  "standard",
                  "custom"
                ],
                "example": "standard"
              },
              "object": {
                "type": "string",
                "example": "contact"
              },
              "connection_id": {
                "type": "string",
                "example": "3217ea51-11c8-43c9-9547-6f197e02e5e4"
              },
              "sync_config_id": {
                "type": "string",
                "example": "3217ea51-11c8-43c9-9547-6f197e02e5e5"
              },
              "paused": {
                "type": "boolean",
                "example": false
              },
              "provider_name": {
                "type": "string",
                "example": "hubspot"
              },
              "customer_id": {
                "type": "string",
                "example": "my-customer-1"
              }
            },
            "required": [
              "id",
              "type",
              "object_type",
              "object",
              "connection_id",
              "sync_config_id",
              "paused",
              "provider_name",
              "customer_id"
            ]
          }
        ]
      },
...
      "create_update_sync_config": {
        "type": "object",
        "properties": {
          "destination_id": {
            "type": "string",
            "example": "6e7baa88-84dd-4dbc-902a-14522c2984eb"
          },
          "provider_id": {
            "type": "string",
            "example": "7f72ec07-e5c1-47fd-8cf5-e71dd13873af"
          },
          "config": {
            "$ref": "#/components/schemas/sync_config_data"
          }
        },
        "required": [
          "destination_id",
          "provider_id",
          "config"
        ]
      },
...
    },
...
  },
...
}